requirements.txt
astroid==2.15.5
cachetools==3.1.1
certifi==2023.5.7
charset-normalizer==3.1.0
dill==0.3.6
google-api-python-client==1.7.2
google-auth==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
httplib2==0.22.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
numpy==1.24.2
oauthlib==3.2.2
platformdirs==3.7.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pylint==2.17.2
pyparsing==3.1.0
python-mpd2==3.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.0
six==1.16.0
tomlkit==0.11.8
uritemplate==3.0.1
urllib3==2.0.3
wrapt==1.15.0
